Mostrar las tareas terminado

+++ src/controllers/tarea_controller.py
@@ -5,7 +5,6 @@ from datetime import datetime
 class TareaController:
     @staticmethod
     async def create_tarea(titulo: str, descripcion: str):
-        # La fecha de creación y el estado de completado se gestionarán automáticamente
         tarea_obj = await Tarea.create(
             titulo=titulo,
             descripcion=descripcion
@@ -13,30 +12,24 @@ class TareaController:
         return tarea_obj
 
     @staticmethod
-    async def read_tarea(tarea_id: int):
-        try:
-            tarea = await Tarea.get(id=tarea_id)
-            return tarea
-        except DoesNotExist:
-            return None
+    async def read_all_tareas():
+     return await Tarea.all()
 
     @staticmethod
     async def update_tarea(tarea_id: int, tarea_data):
-        try:
-            tarea_obj = await Tarea.get(id=tarea_id)
+        tarea_obj = await Tarea.get_or_none(id=tarea_id)
+        if tarea_obj:
             tarea_obj.titulo = tarea_data.titulo
             tarea_obj.descripcion = tarea_data.descripcion
-            tarea_obj.completado = tarea_data.completado
+            tarea_obj.completado = tarea_data.completado  
             await tarea_obj.save()
             return tarea_obj
-        except DoesNotExist:
-            return None
+        return None
 
     @staticmethod
     async def delete_tarea(tarea_id: int):
-        try:
-            tarea = await Tarea.get(id=tarea_id)
+        tarea = await Tarea.get_or_none(id=tarea_id)
+        if tarea:
             await tarea.delete()
             return True
-        except DoesNotExist:
-            return False
+        return False
